Migrate binary search tree
- Create required directories for implementation and tests
- Move files into proper directories
- Migrate binary_search_tree.cpp to header-only implementation
- Add unit tests for binary_search_tree.hpp
- Update README
- Update build script
