+++ C++/README.md
@@ -46,7 +46,7 @@ Implementation of well-known (and some rare) algorithms, in C++.
   * [Edit Distance Problem](include/algorithm/string/edit_distance.hpp) :white_check_mark:
 
 ### Data structures
-* [Binary search tree](include/data_structures/binary_search_tree.cpp)
+* [Binary search tree](include/data_structure/tree/binary_search_tree.hpp) :white_check_mark:
 * [Disjoint-set](include/data_structure/set/disjoint_set.hpp) :white_check_mark:
 * [Singly linked list](include/data_structure/linked_list/singly_linked_list.hpp) :white_check_mark:
